pyproject.toml
[project]
name = "py-crud-02-delicate"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "djangorestframework-simplejwt>=5.5.0",
]
